Refactor server into modular app structure
